19-12-2022 1438

+++ Otros/GloriousRome/src/main/Citizen.java
@@ -20,7 +20,7 @@ public interface Citizen {
 
     public void setFirstName(String text);
 
-    public int getSalary();
+    public float getSalary();
 
     public float getProduction();
